+++ src/deploy.ts
@@ -1,8 +1,8 @@
 import { exit } from 'process';
-import { unlinkSync } from 'fs';
+import fs from 'fs';
 import ora from 'ora';
 import prompts, { PromptObject } from 'prompts';
-import { setup } from './setup';
+import { setupNftSc, setupSftSc } from './setup';
 import {
   deployNftMinterGasLimit,
   outputFileName,
@@ -15,6 +15,7 @@ import {
   deployNftMinterProvenanceHashLabel,
   deployNftMinterTokensLimitPerAddressLabel,
   deployNftMinterImgExtLabel,
+  deploySftMinterGasLimit,
   multiversxExplorer,
   chain,
   nftSCupgradableLabel,
@@ -24,7 +25,8 @@ import {
   deployMetadataInAssetsLabel,
 } from './config';
 import {
-  getDeployTransaction,
+  getDeployNftTransaction,
+  getDeploySftTransaction,
   updateOutputFile,
   getFileContents,
   baseDir,
@@ -37,7 +39,14 @@ const deployNftMinter = async () => {
   const outputFile = getFileContents(outputFileName, { noExitOnError: true });
 
   if (outputFile) {
-    unlinkSync(`${baseDir}/${outputFileName}`);
+    outputFile.nftMinterScAddress = '';
+    outputFile.nftMinterScCollectionSellingPrice = '';
+    outputFile.nftMinterCollectionToken = '';
+    fs.writeFileSync(
+      `${baseDir}/${outputFileName}`,
+      JSON.stringify(outputFile),
+      'utf8'
+    );
   }
 
   const promptsQuestions: PromptObject[] = [
@@ -167,7 +176,7 @@ const deployNftMinter = async () => {
 
   try {
     const { scWasmCode, smartContract, userAccount, signer, provider } =
-      await setup();
+      await setupNftSc();
 
     const {
       deployNftMinterImgCid,
@@ -200,7 +209,7 @@ const deployNftMinter = async () => {
       exit(9);
     }
 
-    const deployTransaction = getDeployTransaction(
+    const deployTransaction = getDeployNftTransaction(
       scWasmCode,
       smartContract,
       deployNftMinterGasLimit,
@@ -252,8 +261,77 @@ const deployNftMinter = async () => {
 
     console.log(`Smart Contract address: ${smartContractAddress}\n`);
     updateOutputFile({
-      scAddress: smartContractAddress,
-      sellingPrice: deployNftMinterSellingPrice,
+      nftScAddress: smartContractAddress,
+      nftSellingPrice: deployNftMinterSellingPrice,
+    });
+  } catch (e) {
+    spinner.stop();
+    console.log((e as Error)?.message);
+  }
+};
+
+export const deploySftMinter = async () => {
+  // Check if there is an old output file
+  const outputFile = getFileContents(outputFileName, { noExitOnError: true });
+
+  if (outputFile) {
+    outputFile.sftMinterScAddress = '';
+    outputFile.sftMinterScCollectionSellingPrice = '';
+    outputFile.sftMinterCollectionToken = '';
+    fs.writeFileSync(
+      `${baseDir}/${outputFileName}`,
+      JSON.stringify(outputFile),
+      'utf8'
+    );
+  }
+
+  const spinner = ora('Processing the transaction...');
+
+  try {
+    const { scWasmCode, smartContract, userAccount, signer, provider } =
+      await setupSftSc();
+
+    await areYouSureAnswer();
+
+    const deployTransaction = getDeploySftTransaction(
+      scWasmCode,
+      smartContract,
+      deploySftMinterGasLimit
+    );
+
+    deployTransaction.setNonce(userAccount.nonce);
+    userAccount.incrementNonce();
+    signer.sign(deployTransaction);
+
+    spinner.start();
+
+    const smartContractAddress = SmartContract.computeAddress(
+      deployTransaction.getSender(),
+      deployTransaction.getNonce()
+    );
+
+    smartContract.setAddress(smartContractAddress);
+
+    await provider.sendTransaction(deployTransaction);
+
+    const watcher = new TransactionWatcher(provider);
+    const transactionOnNetwork = await watcher.awaitCompleted(
+      deployTransaction
+    );
+
+    const txStatus = transactionOnNetwork.status;
+    const txHash = transactionOnNetwork.hash;
+
+    spinner.stop();
+
+    console.log(`\nDeployment transaction executed: ${txStatus}`);
+    console.log(
+      `Deployment tx: ${multiversxExplorer[chain]}/transactions/${txHash}\n`
+    );
+
+    console.log(`Smart Contract address: ${smartContractAddress}\n`);
+    updateOutputFile({
+      sftScAddress: smartContractAddress,
     });
   } catch (e) {
     spinner.stop();
@@ -264,6 +342,7 @@ const deployNftMinter = async () => {
 export const deploy = async (subcommand?: string) => {
   const COMMANDS = {
     nftMinter: 'nft-minter',
+    sftMinter: 'sft-minter',
   };
 
   if (subcommand === '-h' || subcommand === '--help') {
@@ -286,5 +365,8 @@ export const deploy = async (subcommand?: string) => {
 
   if (subcommand === COMMANDS.nftMinter) {
     deployNftMinter();
+  }
+  if (subcommand === COMMANDS.sftMinter) {
+    deploySftMinter();
   }
 };